go/staking/tests: Confirm that node is frozen when slashed
